Add tests for `PackedInts`

+++ src/util/packed/bulk_operation_packed.rs
@@ -359,7 +359,7 @@ where
             let v = values[values_offset];
             values_offset += 1;
             debug_assert!(
-                self.bits_per_value >= PackedInts::unsigned_bits_required(v as u64),
+                self.bits_per_value >= PackedInts::unsigned_bits_required(v),
                 "Value requires more bits than allowed by bits_per_value"
             );
             if (self.bits_per_value as i32) < bits_left {
@@ -443,7 +443,7 @@ where
             let v = values[values_offset];
             values_offset += 1;
             debug_assert!(
-                PackedInts::unsigned_bits_required(v as u64) <= self.bits_per_value,
+                PackedInts::unsigned_bits_required(v as i64) <= self.bits_per_value,
                 "Value requires more bits than allowed by bits_per_value"
             );
             if (self.bits_per_value as i32) < bits_left {
